patch_parse: allow absence of EOL at end of diff
We allow to parse a diff with no EOL at the end of file by loosening the
check for "\n" at end of parsed line when at the last line of the diff
body.

Adding tests for this case as well as for a missing EOL in the middle of
a diff body, which still results in an "invalid patch" error.